SN1-138: renamed files and paths
