--- validators/__init__.py
@@ -1,1 +1,0 @@
-# from .sn1_validator_wrapper import S1ValidatorAPI  # noqa
